# swagger.yaml
...
openapi: '3.0.3'
info:
  title: "SoPekocko API"
  description: "List of all routes availables for the API"
  version: "1.0"
produces: ["application/json"]
servers:
  - url: "http://127.0.0.1:3000"
    description: "Local server test API"

### Components ###
components:
  securitySchemes:
    bearerAuth:           # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  schemas:
    User:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          description: The user email.
          example: "john.doe@mail.tld"
        password:
          type: string
          description: The user password.
          example: "1234"
...
paths:
...
  /api/auth/signup:
    post:
      summary: "Create a user"
      description: "Create a new user and return a message"
      tags: ["Users"]
      requestBody:
        description: "The user to create"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'

      responses:
        201:
          description: 'A message object'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Votre compte a bien été créé."
        401:
          description: 'An error object'
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Error message"
        500:
          description: "Unexpected error"

  # Login
  /api/auth/login:
    post:
      summary: "Login a user"
      description: "Return User object with userId and auth token"
      tags: ["Users"]
      requestBody:
        description: "The user to login"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        200:
          description: 'An object with user ID and auth token'
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    example: "324567UJYGBFDVSTDFYGJ"
                  token:
                    type: string
                    example: "EFRGHJKTR345678UYTHGFD"
        401:
          description: 'An error object'
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Les identifiants fournis ne correspondent pas."
        500:
          description: "Unexpected error"
  
  ### Get user datas in JSON ###
...